requirements.txt
ultralytics>=8.2.0
streamlit>=1.36.0
opencv-python>=4.9.0.80
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
PyYAML>=6.0.1
matplotlib>=3.8.0
pyrebase4>=4.8.0
setuptools>=68
urllib3<2
requests<2.32
gcloud==0.18.3
oauth2client==4.1.3
numpy<2




